pyproject.toml
[tool.poetry]
name = "python-mqtt-examples"
version = "0.1.0"
description = ""
authors = ["Ganesh Chandrasekaran <gchandra@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
paho-mqtt = "^2.1.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
